changed versioning & fixed cleanup

+++ src/janslolcodekernel/kernel.py
@@ -40,4 +40,5 @@ class janslolcodekernel(Kernel):
                }
 
     def do_shutdown(self, restart):
-        shutil.rmtree(workingdir)
+        if os.path.exists(workingdir):
+            shutil.rmtree(workingdir)
